Another round of major reorganization changes...
Renamed "examples" directory to "site" and make it the entire
website instead of just a few loose examples.

All themes now have a <theme-name>.txt file along with their
<theme-name>.css source file. This is a light asciidoc file that provides
a bit of information about the theme. It's used to build the theme
index and on the theme info page.

Added code in "lib/theme.rb" that handles loading theme information
and whatnot.